Shut up some compiler warnings
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@546 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ copyfile.c
@@ -63,8 +63,8 @@ int copyfile(Stream_t *Source, Stream_t *Target)
 				perror("write in copy");
 			else
 				fprintf(stderr,
-					"Short write %lu instead of %d\n",
-					(unsigned long) retw, ret);
+					"Short write %zd instead of %d\n",
+					retw, ret);
 			if(errno == ENOSPC)
 				got_signal = 1;
 			return ret;
